aws_cdk_serverless_dynamo_processor/requirements.txt
-e .
aws.cdk.aws_lambda
aws.cdk.aws_dynamodb
aws.cdk.aws_iam
aws.cdk.aws_lambda_event_sources